# Conflicts:
#	app/Http/Controllers/Admin/Data/PromptController.php
#	app/Models/Prompt/Prompt.php
#	app/Services/PromptService.php
#	app/Services/SubmissionManager.php
#	config/lorekeeper/extension_tracker.php
#	resources/views/admin/prompts/create_edit_prompt.blade.php
#	resources/views/prompts/_prompt_entry.blade.php

+++ app/Services/PromptService.php
@@ -187,6 +187,14 @@ class PromptService extends Service {
                 throw new \Exception('The selected prompt category is invalid.');
             }
 
+            // parent prompt validation
+            if (isset($data['parent_id']) && $data['parent_id'] == 'none') {
+                $data['parent_id'] = null;
+            }
+            if ((isset($data['parent_id']) && $data['parent_id']) && !Prompt::where('id', $data['parent_id'])->exists()) {
+                throw new \Exception('The selected parent prompt is invalid.');
+            }
+
             $data = $this->populateData($data);
 
             $image = null;
@@ -203,7 +211,9 @@ class PromptService extends Service {
                 $data['hide_submissions'] = 0;
             }
 
-            $prompt = Prompt::create(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'staff_only', 'hash', 'level_req']));
+            $prompt = Prompt::create(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'staff_only', 'hash', 'level_req',
+        'parent_id', 'parent_quantity', 'is_details_visible',
+        ]));
 
             if ($image) {
                 $this->handleImage($image, $prompt->imagePath, $prompt->imageFileName);
@@ -249,6 +259,14 @@ class PromptService extends Service {
                 throw new \Exception('That prefix has already been taken.');
             }
 
+            // parent prompt validation
+            if (isset($data['parent_id']) && $data['parent_id'] == 'none') {
+                $data['parent_id'] = null;
+            }
+            if ((isset($data['parent_id']) && $data['parent_id']) && !Prompt::where('id', $data['parent_id'])->exists()) {
+                throw new \Exception('The selected parent prompt is invalid.');
+            }
+
             $data = $this->populateData($data, $prompt);
 
             $image = null;
@@ -263,7 +281,9 @@ class PromptService extends Service {
                 $data['hide_submissions'] = 0;
             }
 
-            $prompt->update(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'staff_only', 'hash', 'level_req']));
+            $prompt->update(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'staff_only', 'hash', 'level_req',
+        'parent_id', 'parent_quantity', 'is_details_visible',
+        ]));
 
             if ($prompt) {
                 $this->handleImage($image, $prompt->imagePath, $prompt->imageFileName);
@@ -295,6 +315,9 @@ class PromptService extends Service {
             // Check first if the category is currently in use
             if (Submission::where('prompt_id', $prompt->id)->exists()) {
                 throw new \Exception('A submission under this prompt exists. Deleting the prompt will break the submission page - consider setting the prompt to be not active instead.');
+            }
+            if (Prompt::where('parent_id', $prompt->id)->exists()) {
+                throw new \Exception('A prompt currently has this prompt as its parent.');
             }
 
             $prompt->rewards()->delete();
@@ -364,6 +387,9 @@ class PromptService extends Service {
         }
         if (!isset($data['level_check'])) {
             $data['level_req'] = null;
+        }
+        if (!isset($data['is_details_visible'])) {
+            $data['is_details_visible'] = 0;
         }
 
         if (isset($data['remove_image'])) {
